refactor: Create datatracker app; move DatatrackerPerson
